infrastructure/cdk/pyproject.toml
[tool.poetry]
name = "energy-comparison-table-cdk"
version = "0.1.0"
description = "CDK app for the Energy Comparison Table"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
aws-cdk-lib = "2.90.0"
constructs = ">=10.0.0,<11.0.0"
pytest = "6.2.5"
ca-cdk-constructs = {git = "https://github.com/citizensadvice/ca-cdk-constructs", rev = "v0.7.0"}

[tool.poetry.group.dev.dependencies]
pre-commit = "2.20.0"

[tool.black]
target-version = ['py311']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
